Fix system.file calls in tests when checking package
Some of the tests included the "inst" directory when attempting to
locate the example data. This works interactively, but not when
running the tests via devtools::check(). Remove the "inst"
subdirectory in the call because the tidyverse overrides it to
work appropriately for interactive use and for installed versions
of a package.