Fix NOTEs about undefined global variables
Due to using dplyr functions, all of the unquoted column names in
the calls were producing NOTEs in R CMD CHECK about undefined
global variables. To fix this, import the .data pronoun from
rlang and prefix all column names with .data$. This is the current
recommended way to handle these notes in the dplyr vignette.