Revert "Add badges for the current branch to the README"
This method of showing the current branch in the readme isn't
working well with merging pull requests on github; it is showing
the badge for the merged branch instead of the badge for the
branch it was merged into because the README file is not
automatically re-rendered. Revert the change and think about if
there's a way to show the current branch more robustly in the
future.

This reverts commit b74148ee1814ddf88d5574c9a8150b3821d0d435.